Rename app to pyfixit2
