+++ pyfixit2/constants.py
