+++ pyfixit2/guide.py
