+++ pyfixit2/test/test_category.py
