+++ pyfixit2/test/test_image.py
@@ -1,7 +1,7 @@
 
 from unittest import TestCase
 
-from pyfixit import Image
+from pyfixit2 import Image
 
 class ImageTest(TestCase):
    @classmethod
